pyproject.toml
[project]
name = "demos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
